+++ src/main/java/org/example/elements/units/DokyuBall.java
@@ -4,7 +4,7 @@ import org.example.elements.Ball;
 import org.example.elements.Bullet;
 
 public class DokyuBall extends Ball {   //弩玉
-    public DokyuBall(float X, float Y, float R, int S, int TYPE) {   //初始化
+    public DokyuBall(float X, float Y, int R, int S, int TYPE) {   //初始化
         super(X, Y, R, S, TYPE);
         speed = 30;
     }
