+++ src/main/java/org/example/elements/units/GuideBall.java
@@ -4,7 +4,7 @@ import org.example.elements.Ball;
 import org.example.elements.atk.MissileBullet;
 
 public class GuideBall extends Ball {   //导玉
-    public GuideBall(float X, float Y, float R, int S, int TYPE) {   //初始化
+    public GuideBall(float X, float Y, int R, int S, int TYPE) {   //初始化
         super(X, Y, R, S, TYPE);
         speed = 180;
     }
