+++ src/main/java/org/example/elements/units/HaneBall.java
@@ -4,7 +4,7 @@ import org.example.elements.Ball;
 import org.example.elements.atk.HaneBullet;
 
 public class HaneBall extends Ball {   //跳玉
-    public HaneBall(float X, float Y, float R, int S, int TYPE) {   //初始化
+    public HaneBall(float X, float Y, int R, int S, int TYPE) {   //初始化
         super(X, Y, R, S, TYPE);
         speed = 50;
     }
