+++ src/main/java/org/example/elements/units/KakuBall.java
@@ -4,7 +4,7 @@ import org.example.elements.Ball;
 import org.example.elements.atk.KakuBullet;
 
 public class KakuBall extends Ball {   //核玉
-    public KakuBall(float X, float Y, float R, int S, int TYPE) {   //初始化
+    public KakuBall(float X, float Y, int R, int S, int TYPE) {   //初始化
         super(X, Y, R, S, TYPE);
         speed = 350;
     }
