+++ src/main/java/org/example/elements/units/ShaBall.java
@@ -4,7 +4,7 @@ import org.example.elements.Ball;
 import org.example.elements.atk.ShaBullet;
 
 public class ShaBall extends Ball {   //射玉
-    public ShaBall(float X, float Y, float R, int S, int TYPE) {   //初始化
+    public ShaBall(float X, float Y, int R, int S, int TYPE) {   //初始化
         super(X, Y, R, S, TYPE);
         speed = 70;
     }
