+++ src/main/java/org/example/elements/units/SwordBall.java
@@ -6,7 +6,7 @@ import org.example.elements.Bullet;
 import org.example.elements.hit.HitsKen;
 
 public class SwordBall extends Ball {     //剑玉
-    public SwordBall(float X, float Y, float R, int S, int TYPE) {
+    public SwordBall(float X, float Y, int R, int S, int TYPE) {
         super(X, Y, R, S, TYPE);
         hp = 30;
         max_hp = 30;
@@ -15,8 +15,7 @@ public class SwordBall extends Ball {     //剑玉
 
     @Override
     public void stepEx(){
-        if (this.cnt == this.speed)
-        {
+        if (this.cnt == this.speed) {
             this.cnt = 0;
             new HitsKen(x, y, rot, side, id, cos_rot, sin_rot);
         }
