+++ src/main/java/org/example/elements/units/UkiBall.java
@@ -4,7 +4,7 @@ import org.example.elements.Ball;
 import org.example.elements.atk.UkiBullet;
 
 public class UkiBall extends Ball {   //浮玉
-    public UkiBall(float X, float Y, float R, int S, int TYPE) {   //初始化
+    public UkiBall(float X, float Y, int R, int S, int TYPE) {   //初始化
         super(X, Y, R, S, TYPE);
         speed = 150;
     }
